+++ src/globus_sdk/experimental/tokenstorage/base.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import abc
 import contextlib
 import os
+import pathlib
 import typing as t
 
 from globus_sdk.services.auth import OAuthTokenResponse
@@ -107,7 +108,20 @@ class FileTokenStorage(TokenStorage, metaclass=abc.ABCMeta):
     files.
     """
 
-    filename: str
+    def __init__(self, filename: pathlib.Path | str, *, namespace: str = "DEFAULT"):
+        """
+        :param filename: the name of the file to write to and read from
+        :param namespace: A user-supplied namespace for partitioning token data
+        """
+        self.filename = str(filename)
+        self._ensure_containing_dir_exists()
+        super().__init__(namespace=namespace)
+
+    def _ensure_containing_dir_exists(self) -> None:
+        """
+        Ensure that the directory containing the given filename exists.
+        """
+        os.makedirs(os.path.dirname(self.filename), exist_ok=True)
 
     def file_exists(self) -> bool:
         """
